+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and to the [CHANGELOG recommendations](http://keepachangelog.com/).
 
 ## ITOU CHANGELOG
 ### [1.4] - (2021-01-25)
-- Tracking session et utilisateur amélioré
+- Tracking session et utilisateur amélioré (listener symfony)
 - Tracking côté back plus étendu (plus robuste, plus fiable)
 
 ### [1.3] - (2021-01-21)